Put core classes into a package.
This is to make organization a little easier in the long run. More work can be theorteically be done, but it is best to move gradually.
