+++ randomizer/constants.py
