+++ randomizer/level_room.py
@@ -1,6 +1,7 @@
 from typing import Dict, List
 from absl import logging
-from constants import Direction, Enemy, Item, Range, RoomNum, RoomType, WallType
+
+from .constants import Direction, Enemy, Item, Range, RoomNum, RoomType, WallType
 
 
 class LevelRoom(object):
